Module Async_js

module Rpc : sig ... end

module Http : sig ... end

module Debug : sig ... end

module Persistent_connection : sig ... end

val sleep : float -> unit Async_kernel.Deferred.t

sleep d is a deferred which becomes determined in d seconds.

val yield : unit -> unit Async_kernel.Deferred.t

yield () returns a deferred that becomes determined after the current cycle completes.

val init : unit -> unit

Initialize the async scheduler

val initialized : unit -> bool

val log_s : Core.Sexp.t -> unit

val log_s_as_string : Core.Sexp.t -> unit