Module Async_js.Debug

val log_s : Core.Sexp.t -> unit

Output a Sexp.t value to the console, e.g. via the message extensions:

log_s [%message "An example" (of_using:Ppx_message.t)]

val log_s_as_string : Core.Sexp.t -> unit

log_s_as_string logs the sexp as a string instead of as a JavaScript object, so that it appears fully expanded in the console.

val any_of_sexp : Core.Sexp.t -> Js_of_ocaml.Js.Unsafe.any

Convert a sexp expression into a JavaScript object. This can be used to output OCaml values to the console.