Module Async_js.Rpc

include module type of struct include Async_rpc_kernel.Rpc end

module Description : sig ... end

module On_exception : sig ... end

When your implementation raises an exception, that exception might happen before a value is returned or after a value is returned. The latter kind of exception is what the ~rest flag to Monitor.try_with is dealing with.

module Implementation : sig ... end

A 'connection_state Implementation.t is something that knows how to respond to one query, given a 'connection_state. That is, you can create a 'connection_state Implementation.t by providing a function which takes a query *and* a 'connection_state and provides a response.

module Implementations : sig ... end

A 'connection_state Implementations.t is something that knows how to respond to many different queries. It is conceptually a package of 'connection_state Implementation.ts.

module Transport : sig ... end

RPC transport layer

module Rpc : sig ... end

module Pipe_close_reason : sig ... end

module Pipe_rpc : sig ... end

module State_rpc : sig ... end

A state rpc is an easy way for two processes to synchronize a data structure by sending updates over the wire. It's basically a pipe rpc that sends/receives an initial state of the data structure, and then updates, and applies the updates under the covers.

module One_way : sig ... end

An RPC that has no response. Error handling is trickier here than it is for RPCs with responses, as there is no reasonable place to put an error if something goes wrong. Because of this, in the event of an error such as dispatching to an unimplemented RPC, the connection will be shut down. Similarly, if the implementation raises an exception, the connection will be shut down.

module Any : sig ... end

module Stable : sig ... end

module Connection : sig ... end