Module Pipe_rpc.Pipe_response

The output type of the f passed to dispatch_iter. This is analagous to a simple unit Deferred.t, with Continue being like Deferred.unit, but it is made explicit when no waiting should occur.

type t =

| Continue

| Wait of unit Async_kernel.Deferred.t